requirements.txt
asgiref==3.6.0
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
constantly==15.1.0
cryptography==40.0.1
daphne==4.0.0
Django==4.1.7
django-admin-interface==0.24.2
django-colorfield==0.8.0
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Pillow==9.5.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.1.1
pytz==2023.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
uni-slugify==0.1.4
Unidecode==1.3.6
zope.interface==6.0
